+++ src/reactpy/web/utils.py
@@ -12,7 +12,7 @@ def module_name_suffix(name: str) -> str:
     if name.startswith("@"):
         name = name[1:]
     head, _, tail = name.partition("@")  # handle version identifier
-    version, _, tail = tail.partition("/")  # get section after version
+    _, _, tail = tail.partition("/")  # get section after version
     return PurePosixPath(tail or head).suffix or ".js"