+++ tests/conftest.py
@@ -42,7 +42,14 @@ def install_playwright():
 
 @pytest.fixture(autouse=True, scope="session")
 def rebuild():
-    subprocess.run(["hatch", "build", "-t", "wheel"], check=True)  # noqa: S607, S603
+    # When running inside `hatch test`, the `HATCH_ENV_ACTIVE` environment variable
+    # is set. If we try to run `hatch build` with this variable set, Hatch will
+    # complain that the current environment is not a builder environment.
+    # To fix this, we remove `HATCH_ENV_ACTIVE` from the environment variables
+    # passed to the subprocess.
+    env = os.environ.copy()
+    env.pop("HATCH_ENV_ACTIVE", None)
+    subprocess.run(["hatch", "build", "-t", "wheel"], check=True, env=env)  # noqa: S607, S603
 
 
 @pytest.fixture(autouse=True, scope="function")